refactor: rename test files to reflect new package structure for edit note functionality
